requirements/ec2.txt
boto3
duckdb
lightgbm
mlflow
numpy
pandas
psycopg2-binary
pyarrow
scikit-learn
xgboost
